[refactor/241130] Refactor diary-related types and components for improved code organization
- Moved DiaryEntry and DiaryDocument type definitions to a centralized types file for better maintainability.
- Updated imports in page and API files to use the new types.
- Removed unused code related to music player from DiaryPage component to streamline functionality.
- Refactored Calendar component to utilize the new CalendarProps type.

This change enhances code clarity and promotes reusability across the application.
